+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `actions.toggle_checkbox` can be mapped in visual mode.
 - `actions.add_property` to add key values to frontmatter.
 
+### Changed
+
+- No longer throw frontmatter warnings with searches that loading notes.
+
 ## [v3.15.5](https://github.com/obsidian-nvim/obsidian.nvim/releases/tag/v3.15.5) - 2026-01-28
 
 ### Added